Project Pat/comp6000-2024-taskmanagementproductivity/Backend3/django_user_registration/user_registration/requirements.txt
Django==5.1.7
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
